New Models et al.
Added Camera Controller
Added signage models
Added new race ships, incl. tuning
- GT Racer
- Formula Racer

+++ Assets/Scripts/SpacecraftData.cs
@@ -6,17 +6,17 @@ using UnityEngine;
 public class SpacecraftData : ScriptableObject
 {
     [Range(0,10), Tooltip("How fast this craft can go. Higher is faster")]
-    public float Speed = 5;
+    public int Speed = 5;
 
     [Range(0, 10), Tooltip("How quickly this craft accelerates to top speed. Higher is quicker")]
-    public float Acceleration = 5;
+    public int Acceleration = 5;
 
     [Range(0, 10), Tooltip("How much boost this craft gets.")]
-    public float Boost = 5;
+    public int Boost = 5;
 
     [Range(0, 10), Tooltip("How fast this craft turns. Higher = faster turns")]
-    public float TurnRate = 5;
+    public int TurnRate = 5;
 
     [Range(0, 10), Tooltip("How floaty this craft is. 1 = tight, 10 = floaty")]
-    public float Drift = 5;
+    public int Drift = 5;
 }
